requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# HTTP Requests
requests==2.31.0

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
fake-useragent==1.4.0

# RSS Feed Parsing - Updated for Python 3.13 compatibility
feedparser==6.0.11

# Date/Time Handling
python-dateutil==2.8.2

# Environment Variables
python-dotenv==1.0.0

# OpenAI for AI Summaries
openai>=1.40.0

# Data Processing - Updated to compatible version
lxml==5.3.0

# File Processing and Analysis - Updated for Python 3.13 compatibility
Pillow>=10.4.0
pytesseract==0.3.10
PyPDF2==3.0.1
pdfplumber==0.9.0
python-magic==0.4.27

# Python 3.13 compatibility fix
legacy-cgi==2.6.3

# PDF generation fallbacks (server-side HTML->PDF)
WeasyPrint==62.3
xhtml2pdf==0.2.15

# Push Notifications
PyAPNs2==2.0.0

# Security & Authentication
flask-jwt-extended==4.6.0
PyJWT==2.8.0
cryptography==42.0.0
flask-limiter==3.5.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# Redis (for rate limiting and token blacklist)
redis==5.0.1

# Security Monitoring
sentry-sdk[flask]==1.40.0

# Password Hashing
bcrypt==4.1.2

# TOTP for MFA
pyotp==2.9.0
qrcode==7.4.2
